add makefile and gitignore

+++ server/server.c
@@ -3,8 +3,8 @@
 #include <errno.h>
 #include <string.h>
 
-#include "server2.h"
-#include "client2.h"
+#include "server.h"
+#include "client.h"
 
 static void init(void)
 {
@@ -71,7 +71,7 @@ static void app(void)
       {
          /* new client */
          SOCKADDR_IN csin = { 0 };
-         size_t sinsize = sizeof csin;
+         unsigned int sinsize = sizeof csin;
          int csock = accept(sock, (SOCKADDR *)&csin, &sinsize);
          if(csock == SOCKET_ERROR)
          {
